Turn type into an enum since there are only two valid values

+++ src/main/kotlin/com/spectralogic/rioclient/General.kt
@@ -6,8 +6,8 @@
 package com.spectralogic.rioclient
 
 data class PageInfo(
-    val number: Int,
-    val pageSize: Int,
+    val number: Long,
+    val pageSize: Long,
     val totalPages: Long,
     val totalItems: Long = 0L
 )
